+++ database/migrations/2023_12_18_101558_create_races_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("name");
             $table->string("description");
-            $table->integer("experience_cost");
+            $table->float("experience_cost", 3,0);
             $table->integer("hit_points");
             $table->foreignId('book_id')->constrained('books');
             $table->timestamps();
